Add a checkbox to remove the upload directory.

+++ tests/test3.nim
@@ -15,6 +15,8 @@ proc handler(req: Request) {.async.} =
   File 2: <input type="file" name="testfile-2" accept="text/*"><br /><br />
   Input 1: <input type="text" name="testfield-1" value="Test"><br /><br />
   Input 2: <input type="text" name="testfield-2" value="Test"><br /><br />
+  <input type="checkbox" name="remove_upload_dir" value="yes" checked> Remove Upload Directory<br />
+  <br />
   <input type="submit">
 </form>
 <br>
@@ -49,8 +51,11 @@ $1
       html.add("</ul>")
       html.add("Upload Directory: $1" % uploadDir)
 
+      if httpbody.formdata.hasKey("remove_upload_dir") and httpbody.formdata["remove_upload_dir"] == "yes":
+        removeDir(uploadDir)
+        html.add(" (Removed)")
+
     await req.respond(Http200, htmlpage % html)
-    removeDir(uploadDir)
 
   await req.respond(Http200, htmlpage % "No data!")
